compile unit for pusher stage
