+++ include/picongpu/simulation/stage/ParticlePush.hpp
@@ -21,8 +21,6 @@
 
 #pragma once
 
-#include "picongpu/particles/ParticlesFunctors.hpp"
-
 #include <pmacc/eventSystem/Manager.hpp>
 
 #include <cstdint>
@@ -42,14 +40,7 @@ namespace picongpu
                  * @param step index of time iteration
                  * @param[out] commEvent particle communication event
                  */
-                void operator()(uint32_t const step, pmacc::EventTask& commEvent) const
-                {
-                    pmacc::EventTask initEvent = eventSystem::getTransactionEvent();
-                    pmacc::EventTask updateEvent;
-                    particles::PushAllSpecies pushAllSpecies;
-                    pushAllSpecies(step, initEvent, updateEvent, commEvent);
-                    eventSystem::setTransactionEvent(updateEvent);
-                }
+                void operator()(uint32_t const step, pmacc::EventTask& commEvent) const;
             };
 
         } // namespace stage
